requirements.txt
PyYAML==3.10
mockfs==1.0.1
